[SvelteKit dev bug] Attempt to fix 1
Will commit and revert right after as a reminder. It feels wrong to do anything with the code just because dev mode fails on one test scene: gltf-to-comps -> problematic Mesh and Empty component imports in the SvelthreeGTLTF class + dev mode doesn' like this test scene, all other test scenes are ok. If I build with the static or the vercel adapter, everything's ok, so there must be some error in the project setup concerning module resolution etc. It's a pain.