`WebGLRenderer` rename `schedule_render` to `schedule_render_auto`
This is needed because we want to introduce a new `WebGLRenderer` method `schedule_render` which will allow users to schedule a render on next AnimationFrame manually (once). Beside that, it feels more consistent, because we have `schedule_render_always` for mode `always` and and now also `schedule_render_auto` for mode `auto` (`schedule_render` was used in mode `auto` logic only).