`types-extra.d.ts` add new types `SvelthreeComponentShadowDOMRoot` and `SvelthreeComponentShadowDOMChild` (#100)
We want to distinguish between components which are generating shadow DOM elements as children only and the `Canvas` component which is generating the `ShadowRoot` element and also sort those components which are not generating any shadow DOM elements.

Also: `Canvas` has a different named getter for retrieving the shadow root element: `get_shadow_root_el()`.