`SvelthreeInteraction` Fix wrong `block` value on listeners update
I stumbled upon this while changing `on_<event_name>` (removed) to `comp.on(...)` in a test which sets interaction programmatically. `comp.on(...)`  function lets `SvelthreeInteraction` component update listeners if callbacks have changed, so `block` was falsely set to `true` even though the object is interactive, so cursor didn't change to pointer when hovering over it, interaction chain was working as expected.