`SvelthreeInteraction` Introduce class `PointerEventManager`
Use a class instead of Closures to manage Pointer Events. This is more memory friendly and probably faster especially when using a lot of interactive objects. Also, this way I was able to move a significant amount of code out of the component and the code is easier to grasp, especially because e.g. `m_pointer.shadowdom_dispatch.overout(evt)` now tells that the event is being dispatched via shadow DOM, as opposed to former `check_pointer_overout(evt)`, which is shorter, but doesn't doesn't include that important information quickly.

This should be done for all Event types.