`SvelthreeGLTF` generate `dom_target` (simple HTMLElement) if a ShadowDOM-Element is not available as target
When creating components via `new` keyword / constructor, `target` has to be set to an `HTMLElement` of some kind. In case `Canvas.useShadowDom` prop was set to `false`, `parent_component.get_shadow_dom_el()` will return `undefined`. In this case we create a simple (not appended) `HTMLElement` to use it as `target` value.