`utils\interaction\event_utils.ts` Refactor `register_event` function, optional `canvas_component` argument
Make `canvas_component` argument optional. Register event in `canvas_component` only if it has been provided. We need this to handle the special case with `KeyboardEvent`s, where the Listener is being added directly to the ShadowDOM-Element (_when using the modifier 'self' -> will be renamed_). In this case we don't have to add a Listener to the `<canvas>`-DOM-Element.